Refactoring
Re-added functionality to add a marker on the map canvas upon clicking
