--- connect.php
@@ -1,5 +1,0 @@
-<?php
-$con = mysqli_connect("localhost", "root", "", "test");
-
-  
-
