+++ js/ShowText.js
@@ -1,0 +1,16 @@
+function showText() {
+    document.getElementById("demo").innerHTML = "<p>FAQ <br><br>" +
+        "How do I use this map?<br> <br>" +
+        "<i>When you find a task with a good reward, locate the corresponding stop on the map (indicated by a google map marker) and use the dropdown menu to make the proper selection and click submit. Stops that you have visited or that aren't worth marking you can mark checked.</i> <br><br> Why doesn't this show all of the quests and rewards?<br><br>" +
+        " <i>To speed up the process of submitting quests only rewards or quests that are worth hunting for are included in this map</i><br><br>" +
+
+        "       Built with <br> HTML / JavaScript / AJAX / PHP / CSS <br> " +
+
+        "     <b>Resources<br></b> " +
+        '    <i>Google Maps JavaScript API <br> ' +
+        'Stack Overflow <br> ' +
+        'PokemonGoHub.net <br> ' +
+        'The Silph Road <br> ' +
+        'Latlong.net' +
+        '</P></i>';
+}
