+++ php/connect.php
@@ -1,0 +1,2 @@
+<?php
+$con = mysqli_connect("localhost", "root", "", "test"); ?>
