+++ php/phpsqlinfo_dbinfo.php
@@ -1,0 +1,5 @@
+<?php
+$username="root";
+$password="";
+$database="test";
+?>
