--- phpsqlinfo_dbinfo.php
@@ -1,5 +1,0 @@
-<?php
-$username="root";
-$password="";
-$database="test";
-?>
